Simplify rotX to vecX
Revert this change if I'm wrong, but I believe we can get away with
understanding X as a vector in ℝ³ rather than a 3 × 1 matrix, despite
the authors emphasizing the isomorphism ℝ³ ≅ ℝ^{3×1} in Definition 2.
In particular, I think Lemma 14 is best proved with just the vector
norm rather than the operator norm. I couldn't find any uses of Lemma
13 that seemed to rely on the interpretation of X as a matrix.